Fix rubocop offenses
[#149756079](https://www.pivotaltracker.com/story/show/149756079)

+++ src/bosh-director/lib/bosh/director/api/link_manager.rb
@@ -8,20 +8,18 @@ module Bosh::Director
         @links_manager = nil
       end
 
-      #TODO Links: if we want to track the user link_create task, what to do with the username
+      # TODO: Links: if we want to track the user link_create task, what to do with the username
       def create_link(username, json_payload)
         validate_link_payload(json_payload)
         consumer_data = json_payload['link_consumer']
 
         provider_intent = find_provider_intent(json_payload['link_provider_id'])
-        if provider_intent.nil?
-          raise "Invalid link_provider_id: #{json_payload["link_provider_id"]}"
-        end
+        raise "Invalid link_provider_id: #{json_payload['link_provider_id']}" if provider_intent.nil?
 
         validate_consumer_network(provider_intent, json_payload['network'], consumer_data['owner_object_name']) unless json_payload['network'].nil?
 
         @links_manager = Bosh::Director::Links::LinksManager.new(Bosh::Director::Config.logger, Bosh::Director::Config.event_log, provider_intent.serial_id)
-        provider = provider_intent.link_provider #find_provider_by_id(provider_intent.link_provider_id)
+        provider = provider_intent.link_provider # find_provider_by_id(provider_intent.link_provider_id)
 
         consumer = @links_manager.find_or_create_consumer(
           deployment_model: provider.deployment,
@@ -40,9 +38,7 @@ module Bosh::Director
 
       def delete_link(username, link_id)
         link = find_link(link_id)
-        if link.nil?
-          raise Bosh::Director::LinkLookupError, "Invalid link id: #{link_id}"
-        end
+        raise Bosh::Director::LinkLookupError, "Invalid link id: #{link_id}" if link.nil?
 
         validate_link_external_type(link)
 
@@ -75,18 +71,16 @@ module Bosh::Director
           consumer = consumer_intent.link_consumer
           consumer_intent.destroy
 
-          if consumer.intents.nil? || consumer.intents.empty?
-            consumer.destroy
-          end
+          consumer.destroy if consumer.intents.nil? || consumer.intents.empty?
         end
       end
 
-      # TODO Links: all provider and consumer related functions should be moved to respective managers/controller?
+      # TODO: Links: all provider and consumer related functions should be moved to respective managers/controller?
 
       def validate_link_payload(json_payload)
-        if  json_payload['link_provider_id'].nil? || json_payload['link_provider_id'] == ''
+        if json_payload['link_provider_id'].nil? || json_payload['link_provider_id'] == ''
           raise 'Invalid request: `link_provider_id` must be provided'
-        elsif !(json_payload['link_provider_id'].is_a?(String))
+        elsif !json_payload['link_provider_id'].is_a?(String)
           raise 'Invalid request: `link_provider_id` must be a String'
         elsif json_payload['link_consumer'].nil?
           raise 'Invalid request: `link_consumer` section must be defined'
@@ -98,7 +92,7 @@ module Bosh::Director
       end
 
       def validate_consumer_network(provider_intent, consumer_network, consumer_name)
-        # TODO Links: discuss about possibility of value of content being empty; will cause nil class error
+        # TODO: Links: discuss about possibility of value of content being empty; will cause nil class error
         content = provider_intent.content
         provider_intent_networks = JSON.parse(content)['networks']
 
@@ -109,9 +103,7 @@ module Bosh::Director
 
       def validate_link_external_type(link)
         consumer = link.link_consumer_intent.link_consumer
-        if consumer.type != @external_type
-          raise Bosh::Director::LinkNotExternalError, 'Error deleting link: not a external link'
-        end
+        raise Bosh::Director::LinkNotExternalError, 'Error deleting link: not a external link' if consumer.type != @external_type
       end
 
       def filter_content_and_create_link(consumer_intent)
